# Conflicts:
#	src/manage-fonts/manage-fonts.css

+++ src/manage-fonts/font-family.js
@@ -1,28 +1,35 @@
-import { useState } from 'react';
+import { useContext } from '@wordpress/element';
 import { Button, Icon } from '@wordpress/components';
 import FontFace from "./font-face";
+import { ManageFontsContext } from '../fonts-context';
 
-const { __ } = wp.i18n;
+const { __, _n } = wp.i18n;
 function FontFamily ( { fontFamily, fontFamilyIndex, deleteFontFamily, deleteFontFace } ) {
 
-    const [isOpen, setIsOpen] = useState(true);
+    const { familiesOpen, handleToggleFamily } = useContext(ManageFontsContext);
+    const isOpen = familiesOpen.includes( fontFamily.name || fontFamily.fontFamily );
 
     const toggleIsOpen = () => {
-        setIsOpen(!isOpen);
+        handleToggleFamily( fontFamily.name || fontFamily.fontFamily );
     }
-
-    const hasFontFaces = !!fontFamily.fontFace && !!fontFamily.fontFace.length;
 
     if ( fontFamily.shouldBeRemoved ) {
         return null;
     }
 
+    const hasFontFaces = !!fontFamily.fontFace && !!fontFamily.fontFace.length;
+
     return (
         <table className="wp-list-table widefat table-view-list">
-            <thead>
+            <thead onClick={toggleIsOpen}>
                 <tr>
                     <td className="font-family-head">
-                        <div><strong>{fontFamily.name || fontFamily.fontFamily}</strong></div>
+                        <div>
+                            <strong>{fontFamily.name || fontFamily.fontFamily}</strong>
+                            { hasFontFaces &&
+                                <span className="variants-count"> ( { fontFamily.fontFace.length } { _n( "Variant", "Variants",  fontFamily.fontFace.length, "create-block-theme" ) } )</span>
+                            }
+                        </div>
                         <div>
                             <Button
                                 variant="tertiary"
@@ -53,21 +60,28 @@ function FontFamily ( { fontFamily, fontFamilyIndex, deleteFontFamily, deleteFon
                                 </tr>
                             </thead>
                             <tbody>
-                                { hasFontFaces && fontFamily.fontFace.map((fontFace, i) => (
-                                    <FontFace
-                                        { ...fontFace }
-                                        fontFamilyIndex={fontFamilyIndex}
-                                        fontFaceIndex={i}
-                                        key={`fontface${i}`}
-                                        deleteFontFace={
-                                            () => deleteFontFace(fontFamilyIndex, i)
-                                        }
-                                    />
-                                )) }
+                                { hasFontFaces && fontFamily.fontFace.map((fontFace, i) => {
+                                    if ( fontFace.shouldBeRemoved ) {
+                                        return null;
+                                    }
+                                    return (
+                                        <FontFace
+                                            { ...fontFace }
+                                            fontFamilyIndex={fontFamilyIndex}
+                                            fontFaceIndex={i}
+                                            key={`fontface${i}`}
+                                            deleteFontFace={
+                                                () => deleteFontFace(fontFamilyIndex, i)
+                                            }
+                                            isFamilyOpen={isOpen}
+                                        />
+                                    )
+                                }) }
                                 {
                                     ! hasFontFaces && fontFamily.fontFamily &&
                                     <FontFace
                                         { ...fontFamily }
+                                        isFamilyOpen={isOpen}
                                     />
                                 }
                             </tbody>
